services/speech-to-text/requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
werkzeug==3.0.1
numpy<2.0.0

# Speech-to-Text - Whisper
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0

# gRPC dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
